python-rtc/requirements.txt
videosdk=0.0.2
python-dotenv=1.0.1
asyncio=3.4.3